# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      error:
        type: string
    type: object
  users.CreateUserRequest:
    properties:
      dataOfBirthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  users.CreateUserResponse:
    properties:
      dataOfBirthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        type: string
      lastName:
        type: string
    type: object
  users.GetAllUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/users.UserResponse'
        type: array
    type: object
  users.Message:
    properties:
      message:
        type: string
    type: object
  users.UpdateUserRequest:
    properties:
      Id:
        type: string
      LastName:
        type: string
      dataOfBirthday:
        type: string
      firstName:
        type: string
    type: object
  users.UserResponse:
    properties:
      dataOfBirthday:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: API-Gateway service
  version: "1.0"
paths:
...
  /api/user:
    delete:
      description: Deletes the user based on JWT claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete User
      tags:
      - User
    get:
      description: Retrieves the user profile based on JWT claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates user information based on the provided request body and
        JWT claims
      parameters:
      - description: Update User Request
        in: body
        name: UpdateUserRequest
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update User
      tags:
      - User
  /api/user/all:
    get:
      description: Retrieves a list of users based on query parameters
      parameters:
      - description: Limit the number of users returned
        in: query
        name: limit
        type: string
      - description: Offset for pagination
        in: query
        name: offset
        type: string
      - description: Filter by first name
        in: query
        name: firstName
        type: string
      - description: Filter by gender
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.GetAllUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get All Users
      tags:
      - User
  /api/user/create:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: Create User Request
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create User
      tags:
      - User
...
schemes:
- http
swagger: "2.0"